chore(feedback dashboard): refactored, renamed, and improved refreshing
