+++ app/client/src/features/events/LiveFeedbackPrompts/LiveFeedbackPrompt/useLiveFeedbackPrompts.tsx
@@ -36,19 +36,14 @@ const USE_LIVE_FEEDBACK_PROMPTS = graphql`
 
 export interface Props {
     fragmentRef: useLiveFeedbackPromptsFragment$key;
-    isModalOpen: boolean;
-    isShareResultsOpen: boolean;
 }
 
-export function useLiveFeedbackPrompts({ fragmentRef, isModalOpen, isShareResultsOpen }: Props) {
+export function useLiveFeedbackPrompts({ fragmentRef }: Props) {
     const [data, refetch] = useRefetchableFragment(USE_LIVE_FEEDBACK_PROMPTS, fragmentRef);
     const { liveFeedbackPrompts } = data;
 
     const REFETCH_INTERVAL = 20000; // 20 seconds
     const refresh = React.useCallback(() => {
-        // if the modal is open, don't refetch (Ensures secondary modal doesn't flash)
-        if (isModalOpen) return;
-
         const endCursor = data.liveFeedbackPrompts?.pageInfo?.endCursor;
 
         // Prepare variables conditionally
@@ -62,16 +57,8 @@ export function useLiveFeedbackPrompts({ fragmentRef, isModalOpen, isShareResult
         }
 
         refetch(variables, { fetchPolicy: 'store-and-network' });
-    }, [isModalOpen, refetch, data.liveFeedbackPrompts?.pageInfo?.endCursor]);
-    const { pauseRefresh, resumeRefresh } = useRefresh({ refreshInterval: REFETCH_INTERVAL, callback: refresh });
-
-    React.useEffect(() => {
-        if (isModalOpen || !isShareResultsOpen) {
-            pauseRefresh();
-        } else {
-            resumeRefresh();
-        }
-    }, [isModalOpen, isShareResultsOpen, pauseRefresh, resumeRefresh]);
+    }, [refetch, data.liveFeedbackPrompts?.pageInfo?.endCursor]);
+    useRefresh({ refreshInterval: REFETCH_INTERVAL, callback: refresh });
 
     const promptsList = React.useMemo(
         () =>
@@ -86,5 +73,5 @@ export function useLiveFeedbackPrompts({ fragmentRef, isModalOpen, isShareResult
         [data.liveFeedbackPrompts?.__id]
     );
 
-    return { prompts: promptsList, connections, pauseRefresh, resumeRefresh, refresh };
+    return { prompts: promptsList, connections, refresh };
 }
